more dynamic calendar
now weekdays are displayed, and we can start on any weekday, and it will be ofset accordingly (ex wednesday will start on the 3rd tile, friday on the 5th, and so on) All months are 30 days in this game. Graphics are still placeholder, proof-of-concept UI for now. Next, indicate which day we´re on, and bring up more information when day is clicked.