inventory items
inventory_ui.gd now takes inventory data from inventory_data.json and dynamically displays sprites corresponding to that data in the inventory panel, according to which tab is active.

Next is RMB to display item information, and LMB to select and use > set custom cursor to inventory item sprite, until clicked on an NPC or object.