Allow for VN type games - part 1
It was my intention from the beginning to have this engine support both 2d adventure games/3D Adventure Games and Visual novels (and to be able to switch between the three on a per scene basis). All scenes contain the information needed to make this switch pretty much automatic, but there still need to be some checks in the engine. So now, we determine if we´re doing a 3d adventure game or Visual Novel-style game, by checking for type of the first node in the scene loaded (Area/Area2D)
. We´ve also added a global.gameType variable that we can use to check type with from other scripts.

This is just the first step. For now, we don´t load the player scene unless scene is 3D (since player coords are Vector3), but in the future we will be able load a 2d player too. However 3d adventure games and Visual Novels are prioritized, because that´s what I´m currently
working on.

TODO: code currently assumes 3d meshes when placing Player, NPCs and Objects (Using a Vector3) - need to
 allow for Vector2 position. Also, probably something else I can´t think about but will propably bite me in the ass when I get there - but we´ll handle that then :)