Use items from inventory - part 1
When clicking an item in your inventory, the mouse cursor now changes to represent the item you picked. When hovering over an NPC you will now be asked if you want to give the item the NPC.

Youstill can´t actually give the item to the character (TODO!) but at least of the preparatory work is done.

Rightclicking returns the cursor to normal.

TODO: make sure that the item name is referenced when hovering over NPC/object.
TODO: Ability to actually give the item to the NPC/object and get a response
TODO: Once given, the object needs to be removed from your inventory.