Use items from inventory - part 1b
clicking on an NPC now changes cursor back to default. I wanted to make this remove the item from the inventory, which exposed a design flaw with inventory_data.json.

I´m going have to assign every item a number for easier iteration. Waiting with that, since it means I´ll have to change code for displaying the inventory items.