Grab screenshot with thumbnail
Created a global.grab_screen() func with the purpose being saving screenshots for save games (for now). Not sure how to do this when pressing save game in game settings menu, as it will catch the game settings menu. So I cache an image before game settings opens. Right now this causes the transition to game settings to pause momentarily. Maybe solved by just capturing screenshot when opening menu, and not resizing and save thumb until clicking on save?