Game settings menu - save game screenshot thumbnail
Warning - ugly code :P

Instead of showing 6 save games right away, the first save slot is marked by a + indicating we need to click it to save the game (right now, we´re only saving a screenshot and displaying the thumb, but whatever). When clicking the +, it´s replaced by a thumbnail image, and the next slot get´s the +. Only works for the first slot for now, but pretty straight forward to implement to the rest now, and then we need to work on similar paging functionality as in the phone gallery.

Took me awhile to get right, and for now I´m just happy it works. Will clean up at a later stage.

Also, commented out scene transition code for now, as it´s not working properly. Uses cached previous screenshot for some reason.