work on save slots - part 1
Further work on save slot system. Created a new func save_to_slot(id) to handle the slots dynamically, so every time you click on a plus, it will "save" a thumbnail to that slot, and create a plus in the next slot.

Since paging isn´t implemented yet, the game will crash when clicking on the final slot, but we´re getting there.

Decided to take a rest today, so this (actually surprisingly quick) hack is the exception, since it was fairly easy to implement ;P