AnimationPlayer experiments
Continued playing around with Jayanam´s character to learn about how to control the AnimationPlayer in Godot,
to add run and idle animations. This forced me to reviit the player.gd code fo the first time in months
(that code scares me!), and actually allowed me to streamline the code a bit :) Now the player will use the
idle animation when standing still/stopping and the run animation when moving. Obviously this is just a test
with someone elses animations, but it´s a nice first prototype code. Need to add check so clicking when moving
doesn´t restart animation.