Fixes, tweaks and json code refactoring
clicking on or hovering an NPC while in any overlay menu will no longer accidentally trigger NPC dialogue, or hover texts. Flag was missing.

When hovering over NPC with item, both item and NPC name is displayed now: "Give rose to Ellie?"

location jsons now have a "default" section, which is used if current day is not found. This is useful for locations that don´t need a special setting for every day of the week, to keep the filesize down, and make editing easier. Touched files in both global.gd and dialogue.gd, and I hope I didn´t break any section of the dialogue I have not yet used..