Don´t move dummy
Fixed a longstanding bug where the player would start moving when exiting a menu or swicthing scenes. Solution in the end isn´t pretty, but it works. Created a dummy tween, that runs for 0.5 seconds when exiting a menu, and then sends a signal to ui.gd to set global ui_blocking to false. probably will need some time to remove redundant flags now, but not a priority...