Scene transition adjustments
Before we took a screenshot when entering map_ui and then changed scene as soon as we clicked on the location. That caused some jerky animation in the map ui, so I changed it.

Now we check when the tween that hides the map ui is complete, then take a screenshot and start the transitition. Looks way smoother.