Proper dialogue
Were you tired of the "goto b/goto c" dialogue? Me too, as exciting as it was watching it unfold, a new scenario was needed to display the power of the dialogue/event system, and... this is not it. It is however the first step to creating a scenario spanning more than one scene, invloving some casual drug dealing (if you assume this dialogue is not child friendly, you would be assuming correctly) PG13 at least :)