action reminders/hints
Added an "action" key to the dialogue json, under the "reply" sub.

An action is something you need to do after the dialogue, and will display as a hint/reminder in a thought bubble immediately after the dialogue has closed.

This can of course we added in conjunction with setting up an event.