preparatory work
changed cursor size from 16x16 to 24x24 which is easier to see, and can fit more visual information without being too much in the way.

Added a gear icon for when hovering over UI icons, and an arrow icon, which is used in game settings and UI overlays. Still placeholder art (!)

Created a contacts node for the phone ui, which is not yet connected to contact data. Every contact node has an avatar, id, info, number and a call icon.

TODO: When calling someone the phone UI is miminized and a dialogue is initiated with the phone as avatar.

Tomorrow recommence work on the testing scenario. We need to call Levaun, using the new contact node, and then trigger an event override for later that evening in the school hall.