game events refactor
Ok, this gave me quite some headache, and in the end I had to change the structure of global.gameEvents.

Background: I had originally envisioned the possibility of having more than one event at the same timeofday, so I made an array for each event, which turns out makes everything a mess.

Now gameEvents is pure dictionary and a lot easier to work with. If events conflict in the game, you will be prompted to decide if you want to cancel any other event already booked at same timeofday, and consequences will be regular oneoff events.

No, when you talk to Devaun on the phone, the event triggered is added properly to gameEvents.

I was hoping to get event notifications done by today, let´s see if I have the stamina later. At least we have a simpler system now, that should cause less headache in the future.